requirements_gq.txt


python-dotenv
flask

# LangChain (compatible versions)
langchain
langchain-core
langchain-community
langchain-huggingface

# Pinecone integration
langchain-pinecone

# Google Gemini support
#langchain-google-genai
#google-generativeai

#Groq LLama
langchain-groq==0.1.4

# Embeddings / Vectorization
sentence-transformers
pypdf

langchain-text-splitters
-e .

